Crate fj_plugins

Abstractions for defining your own Fornjot plugins.

The typical workflow is to first define a Model type.

Macros

register_plugin

Declare the function that will be called when a plugin is first initialized.

Structs

MissingArgument

The error returned when a required argument wasn’t provided.

Traits

Context

Contextual information passed to a Model when it is being initialized.

ContextExt

Extension methods for the Context type.

Host

An abstract interface to the Fornjot host.

HostExt

Extension methods to augment the Host API.

Model

A model.

Type Definitions

ModelConstructor

A type-erased function that is called to construct a Model.