Macro fj_plugins::register_plugin

macro_rules! register_plugin {
    ($init:expr) => { ... };
}

Declare the function that will be called when a plugin is first initialized.

This is where you’ll do things like registering a model with the host and so on.

use fj_plugins::{Host, HostExt, Model, Context, PluginMetadata};

fj_plugins::register_plugin!(|host: &mut dyn Host| {
    host.register_model::<MyModel>();
    PluginMetadata::new(env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
});

struct MyModel;

impl Model for MyModel {
    fn from_context(ctx: &dyn Context) -> Result<Self, anyhow::Error>
    where
        Self: Sized,
    {
        todo!("Load arguments from the context and initialize the model");
    }

    fn shape(&self) -> fj::Shape { todo!("Calcualte the model's geometry") }
}