Trait fj_plugins::ContextExt

pub trait ContextExt {
    fn get_required_argument(&self, name: &str) -> Result<&str, MissingArgument>;
    fn parse_argument<T>(&self, name: &str) -> Result<T, ContextError>
    where
        T: FromStr,
        T::Err: Error + Send + Sync + 'static;
}

Extension methods for the Context type.

By splitting these methods out into a separate trait, Context can stay object-safe while allowing convenience methods that use generics.

Required Methods

fn get_required_argument(&self, name: &str) -> Result<&str, MissingArgument>

Get an argument, returning a MissingArgument error if it doesn’t exist.

fn parse_argument<T>(&self, name: &str) -> Result<T, ContextError> where
T: FromStr,
T::Err: Error + Send + Sync + 'static,

Parse an argument from its string representation using FromStr.