Trait fj_plugins::ModelFromContext

pub trait ModelFromContext: Sized {
    fn from_context(ctx: &dyn Context) -> Result<Self, Error>;
}

A Model that can be loaded purely from the Context.

Required Methods

fn from_context(ctx: &dyn Context) -> Result<Self, Error>

Try to initialize this Model using contextual information it has been provided.