Crate fj_plugins

Abstractions for defining your own Fornjot plugins.

The typical workflow is to first define a Model type. This comes with two main methods,

Use the ModelFromContext trait to make the model loadable from the host context, and

Calculate the model’s shape with the Model trait

use fj_plugins::{Model, Context, Error, ModelFromContext};

struct MyModel;

impl ModelFromContext for MyModel {
    fn from_context(ctx: &dyn Context) -> Result<Self, Error>
    where
        Self: Sized,
    {
        todo!("Load arguments from the context and initialize the model");
    }
}

impl Model for MyModel {
    fn shape(&self) -> fj::Shape { todo!("Calcualte the model's geometry") }
}

Once you’ve done that, you can use the register_plugin!() macro to set up the plugin (possibly registering a model) and tell the host some basic information about it.

use fj_plugins::{Host, HostExt, PluginMetadata};

fj_plugins::register_plugin!(|host: &mut dyn Host| {
    host.register_model::<MyModel>();

    Ok(PluginMetadata::new(
        env!("CARGO_PKG_NAME"),
        env!("CARGO_PKG_VERSION"),
    ))
});

Macros

register_plugin

Declare the function that will be called when a plugin is first initialized.

Structs

MissingArgument

The error returned when a required argument wasn’t provided.

PluginMetadata

Information about a particular plugin that can be used by the host for things like introspection and search.

Traits

Context

Contextual information passed to a Model when it is being initialized.

ContextExt

Extension methods for the Context type.

Host

An abstract interface to the Fornjot host.

HostExt

Extension methods to augment the Host API.

Model

A model.

ModelFromContext

A Model that can be loaded purely from the Context.

Type Definitions

Error

The common error type used by this crate.

ModelConstructor

A type-erased function that is called to construct a Model.